AUTHORIZATION

+++ src/main/java/com/urosdragojevic/realbookstore/controller/PersonsController.java
@@ -4,13 +4,16 @@ import com.urosdragojevic.realbookstore.audit.AuditLogger;
 import com.urosdragojevic.realbookstore.domain.Person;
 import com.urosdragojevic.realbookstore.domain.User;
 import com.urosdragojevic.realbookstore.repository.PersonRepository;
+import com.urosdragojevic.realbookstore.repository.RoleRepository;
 import com.urosdragojevic.realbookstore.repository.UserRepository;
 import jakarta.servlet.http.HttpSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -26,13 +29,16 @@ public class PersonsController {
 
     private final PersonRepository personRepository;
     private final UserRepository userRepository;
+    private final RoleRepository roleRepository;
 
-    public PersonsController(PersonRepository personRepository, UserRepository userRepository) {
+    public PersonsController(PersonRepository personRepository, UserRepository userRepository, RoleRepository roleRepository) {
         this.personRepository = personRepository;
         this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
     }
 
     @GetMapping("/persons/{id}")
+    @PreAuthorize("hasAuthority('VIEW_PERSON')")
     public String person(@PathVariable int id, Model model, HttpSession session) {
         String csrf = session.getAttribute("CSRF_TOKEN").toString();
         model.addAttribute("CSRF_TOKEN", session.getAttribute("CSRF_TOKEN"));
@@ -41,6 +47,7 @@ public class PersonsController {
     }
 
     @GetMapping("/myprofile")
+    @PreAuthorize("hasAuthority('VIEW_MY_PROFILE')")
     public String self(Model model, Authentication authentication) {
         User user = (User) authentication.getPrincipal();
         model.addAttribute("person", personRepository.get("" + user.getId()));
@@ -48,21 +55,46 @@ public class PersonsController {
     }
 
     @DeleteMapping("/persons/{id}")
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
     public ResponseEntity<Void> person(@PathVariable int id) {
-        personRepository.delete(id);
-        userRepository.delete(id);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User current = (User)authentication.getPrincipal();
+        if(roleRepository.findByUserId(current.getId()).stream().anyMatch(role ->
+                role.getName().equals("MANAGER") || role.getName().equals("REVIEWER"))) {
+            if (id == current.getId()) {
+                personRepository.delete(id);
+                userRepository.delete(id);
+            }
+        }
+        if(roleRepository.findByUserId(current.getId()).stream().anyMatch(role ->
+                role.getName().equals("ADMIN"))){
+            personRepository.delete(id);
+            userRepository.delete(id);
+        }
 
         return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/update-person")
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
     public String updatePerson(Person person, HttpSession session, @RequestParam("csrfToken") String csrfToken)
             throws AccessDeniedException {
         String csrf = session.getAttribute("CSRF_TOKEN").toString();
         if(!csrf.equals(csrfToken)) {
             throw new AccessDeniedException("Forbidden");
         }
-        personRepository.update(person);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User current = (User) authentication.getPrincipal();
+        if(roleRepository.findByUserId(current.getId()).stream().anyMatch(role ->
+                role.getName().equals("MANAGER") || role.getName().equals("REVIEWER"))) {
+            if (Integer.valueOf(person.getId()) == current.getId()) {
+                personRepository.update(person);
+            }
+        }
+        if(roleRepository.findByUserId(current.getId()).stream().anyMatch(role ->
+                role.getName().equals("ADMIN"))){
+            personRepository.update(person);
+        }
         return "redirect:/persons/" + person.getId();
     }
 
@@ -73,6 +105,7 @@ public class PersonsController {
     }
 
     @GetMapping(value = "/persons/search", produces = "application/json")
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     @ResponseBody
     public List<Person> searchPersons(@RequestParam String searchTerm) throws SQLException {
         return personRepository.search(searchTerm);
